Share grid items rendering advancement between a box and its copies
Without that, when a grid container is copied before its layout (like in the grid
layout), the original grid container doesn’t have the advancement calculated by its
copy.

We should really store this information in the rendering context, and not only for grid,
to correctly follow the fragmentation algorithm.

Fix #2626.